# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cloud Cost Optimization Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CloudAccount {
  id: number;
  /** aws | gcp | azure */
  provider: string;
  name: string;
  accountId: string;
  region: string;
  /** active | inactive */
  status: string;
  monthlyCost: number;
  createdAt: string;
}

export interface CloudAccountInput {
  provider: string;
  name: string;
  accountId: string;
  region: string;
}

export interface CloudResource {
  id: number;
  accountId: number;
  provider: string;
  /** ec2 | rds | s3 | lambda | gce | gcs | bigquery | vm | blob_storage | aks */
  resourceType: string;
  name: string;
  region: string;
  /** active | idle | unused | stopped */
  status: string;
  monthlyCost: number;
  cpuUtilization: number;
  memoryUtilization: number;
  lastActive: string;
  tags: string;
}

export interface CloudResourceUpdate {
  /** @nullable */
  status?: string | null;
  /** @nullable */
  tags?: string | null;
}

export interface Recommendation {
  id: number;
  resourceId: number;
  resourceName: string;
  provider: string;
  /** shutdown | resize_down | delete | schedule | reserved_instance | spot_instance */
  recommendationType: string;
  description: string;
  estimatedMonthlySavings: number;
  /** critical | high | medium | low */
  priority: string;
  /** pending | applied | dismissed */
  status: string;
  createdAt: string;
}

export interface CostAlert {
  id: number;
  /** critical | warning | info */
  severity: string;
  title: string;
  message: string;
  estimatedImpact: number;
  resolved: boolean;
  createdAt: string;
}

export interface CostForecast {
  id: number;
  /** @nullable */
  accountId: number | null;
  forecastDate: string;
  predictedCost: number;
  confidenceLow: number;
  confidenceHigh: number;
}

export interface DashboardSummary {
  totalMonthlyCost: number;
  projectedMonthCost: number;
  potentialMonthlySavings: number;
  totalResources: number;
  idleResources: number;
  unusedResources: number;
  activeAlerts: number;
  pendingRecommendations: number;
  costChangePercent: number;
}

export interface ServiceSpend {
  service: string;
  monthlyCost: number;
  percentage: number;
}

export interface ProviderSpend {
  provider: string;
  monthlyCost: number;
  percentage: number;
  resourceCount: number;
}

export interface DailyCost {
  date: string;
  cost: number;
}

export type ListResourcesParams = {
/**
 * @nullable
 */
accountId?: number | null;
/**
 * @nullable
 */
provider?: string | null;
/**
 * @nullable
 */
status?: string | null;
/**
 * @nullable
 */
resourceType?: string | null;
};

export type ListRecommendationsParams = {
/**
 * @nullable
 */
status?: string | null;
/**
 * @nullable
 */
priority?: string | null;
};

export type ListAlertsParams = {
/**
 * @nullable
 */
severity?: string | null;
/**
 * @nullable
 */
resolved?: boolean | null;
};

export type ListForecastsParams = {
/**
 * @nullable
 */
accountId?: number | null;
/**
 * @nullable
 */
days?: number | null;
};