# lib/api-zod/src/generated/types/cloudAccount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cloud Cost Optimization Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface CloudAccount {
  id: number;
  /** aws | gcp | azure */
  provider: string;
  name: string;
  accountId: string;
  region: string;
  /** active | inactive */
  status: string;
  monthlyCost: number;
  createdAt: string;
}