# lib/api-zod/src/generated/types/cloudResource.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cloud Cost Optimization Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface CloudResource {
  id: number;
  accountId: number;
  provider: string;
  /** ec2 | rds | s3 | lambda | gce | gcs | bigquery | vm | blob_storage | aks */
  resourceType: string;
  name: string;
  region: string;
  /** active | idle | unused | stopped */
  status: string;
  monthlyCost: number;
  cpuUtilization: number;
  memoryUtilization: number;
  lastActive: string;
  tags: string;
}