# lib/api-zod/src/generated/types/costForecast.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cloud Cost Optimization Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface CostForecast {
  id: number;
  /** @nullable */
  accountId: number | null;
  forecastDate: string;
  predictedCost: number;
  confidenceLow: number;
  confidenceHigh: number;
}