# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cloud Cost Optimization Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalMonthlyCost: number;
  projectedMonthCost: number;
  potentialMonthlySavings: number;
  totalResources: number;
  idleResources: number;
  unusedResources: number;
  activeAlerts: number;
  pendingRecommendations: number;
  costChangePercent: number;
}