# lib/api-zod/src/generated/types/providerSpend.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cloud Cost Optimization Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ProviderSpend {
  provider: string;
  monthlyCost: number;
  percentage: number;
  resourceCount: number;
}