# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cloud Cost Optimization Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface Recommendation {
  id: number;
  resourceId: number;
  resourceName: string;
  provider: string;
  /** shutdown | resize_down | delete | schedule | reserved_instance | spot_instance */
  recommendationType: string;
  description: string;
  estimatedMonthlySavings: number;
  /** critical | high | medium | low */
  priority: string;
  /** pending | applied | dismissed */
  status: string;
  createdAt: string;
}