# lib/api-zod/src/generated/types/serviceSpend.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cloud Cost Optimization Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceSpend {
  service: string;
  monthlyCost: number;
  percentage: number;
}